게시물 상세보기 좋아요 수정

+++ SEMI_PROJECT/src/main/java/com/controller/action/CheckLikeAction.java
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import com.dao.PostDao;
-import com.dao.StoryDao;
+import com.dto.MemberDto;
 
 public class CheckLikeAction implements Action {
 
@@ -17,24 +17,24 @@ public class CheckLikeAction implements Action {
 		int post_num = Integer.parseInt(request.getParameter("post_num"));
 		String url = "spring.do?command=postDetail&post_num=" + post_num;
 		HttpSession session = request.getSession();
-		//MeberDto mdto = (MemberDto) sessio.getAttribute("loginAdmin");
-		//if(mdto==null) url = "spring.do?command=login";
-		//else {
-			String userid = "jojo";//((MemberDto) session.getAttribute("loginUser")).getUserid();
+		MemberDto mdto = (MemberDto) session.getAttribute("loginUser");
+		if(mdto==null) url = "spring.do?command=login";
+		else {
+			String userid = mdto.getUserid();
 			PostDao pdao = PostDao.getInstance();
 			int result = pdao.postLikeCheck(post_num, userid);
-			String fileName = "";
+
+			System.out.println("likeResult : " + result);
 			if(result == 0) {
 				pdao.addPostLike(post_num, userid);
-				fileName = "../images/Like.png";
 				result = 1;
 			} else {
 				pdao.deleteLike(post_num, userid);
-				fileName = "../images/beforeLike.png";
 				result = 0;
 			}
 			request.setAttribute("likeResult", result);
-		//}
+			System.out.println("likeResult : " + result);
+		}
 			
 		response.sendRedirect(url);
 	}
